pyproject.toml
# Rename to build stable version
# This is stable version
[build-system]
requires = ["setuptools>=82.0.1"]
build-backend = "setuptools.build_meta"

[project]
name = "je_web_runner"
version = "0.0.67"
authors = [
    { name = "JE-Chen", email = "jechenmailman@gmail.com" },
]
description = "WEB Automation Framework"
requires-python = ">=3.10"
license-files = ["LICENSE"]
dependencies = [
    "selenium>=4.0.0",
    'requests',
    'python-dotenv',
    "webdriver-manager",
    "defusedxml",
    "Pillow",
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Win32 (MS Windows)",
    "Environment :: MacOS X",
    "Environment :: X11 Applications",
    "Operating System :: OS Independent"
]

[project.urls]
Homepage = "https://github.com/Intergration-Automation-Testing/WebRunner"
Documentation = "https://webrunner.readthedocs.io/en/latest/"
Code = "https://github.com/Intergration-Automation-Testing/WebRunner"

[project.readme]
file = "README.md"
content-type = "text/markdown"


[tool.setuptools.packages]
find = { namespaces = false }
